refactor(github-issues): extract shared auth helpers
