+++ crates/tau-github-issues/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 reqwest = { workspace = true }
 chrono = { workspace = true }
 sha2 = { workspace = true }
+serde_json = { workspace = true }
