+++ crates/tau-github-issues/src/lib.rs
@@ -4,6 +4,7 @@
 
 pub mod github_issues_helpers;
 pub mod github_transport_helpers;
+pub mod issue_auth_helpers;
 pub mod issue_comment;
 pub mod issue_demo_index;
 pub mod issue_filter;
